CHG: Fix small bug in logging code.. force values to be (float) CHG: also check that data exists in each buffer, otherwise output nothing

+++ src/viewer/SimViewer.cpp
@@ -360,7 +360,7 @@ void SimViewer::draw()
 
         const float kinEnergy = computeKineticEnergy(*m_rigidBodySystem);
         s_log.pushVal(s_kinEnergyIdx, kinEnergy);
-        s_log.pushVal(s_substepsIdx, m_subSteps);
+        s_log.pushVal(s_substepsIdx, (float)m_subSteps);
     }
 }
